# docs/openapi.json
{
  "components": {
    "schemas": {
      "CashInOutRequest": {
        "description": "Op 11 request: record a cash-drawer adjustment.",
        "properties": {
          "amount": {
            "description": "Amount; must be greater than zero. Sent as a JSON number on the wire.",
            "format": "double",
            "type": "number"
          },
          "cashierId": {
            "description": "Cashier number.",
            "format": "uint32",
            "minimum": 0,
            "type": [
              "integer",
              "null"
            ]
          },
          "description": {
            "default": "",
            "description": "Free-text comment.",
            "type": "string"
          },
          "isCashIn": {
            "description": "`true` for cash-in (deposit), `false` for cash-out (withdrawal).",
            "type": "boolean"
          }
        },
        "required": [
          "amount",
          "isCashIn"
        ],
        "type": "object"
      },
...
      "EmptyResponse": {
        "description": "Marker response for operations that do not return a meaningful payload.\n\nThe wire-level response framing is still parsed (header + zero-byte or `{}` payload), but the\ncaller gets no fields to consume.",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "description": "The shared bridge token (HDM_BRIDGE_TOKEN).",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Local HTTP bridge for the Armenian fiscal cash register (HDM) protocol. Every operation is a `POST /v1/<op>` carrying a uniform JSON envelope `{ connection?, params? }`: `connection` overrides the bridge's configured default device field-by-field, and `params` carries the operation input. On failure every route returns the same error envelope (see `ErrorBody`): `kind` is a stable machine tag and `code` carries the device/spec response code when the device rejected the request. All routes except `/v1/health`, `/v1/info`, and `/v1/openapi.json` require `Authorization: Bearer <token>`.",
    "license": {
      "name": "MIT OR Apache-2.0"
    },
    "title": "HDM Bridge API",
    "version": "0.5.3"
  },
  "openapi": "3.1.0",
  "paths": {
    "/v1/cash": {
      "post": {
        "description": "Register a cash-drawer in or out. Requires connection: host + password + cashier + pin.",
        "operationId": "cashInOut",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "connection": {
                    "$ref": "#/components/schemas/PartialConn"
                  },
                  "params": {
                    "$ref": "#/components/schemas/CashInOutRequest"
                  }
                },
                "required": [
                  "params"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmptyResponse"
                }
              }
            },
            "description": "Operation succeeded."
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Error envelope (4xx/5xx)."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Register a cash-drawer in or out."
      }
    },
    "/v1/datetime": {
      "post": {
        "description": "Get the device date and time. Requires connection: host + password + cashier + pin.",
        "operationId": "dateTime",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "connection": {
                    "$ref": "#/components/schemas/PartialConn"
                  },
                  "params": {
                    "description": "Ignored for this operation."
                  }
                },
                "type": "object"
              }
            }
          },
          "required": false
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DateTimeResponse"
                }
              }
            },
            "description": "Operation succeeded."
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Error envelope (4xx/5xx)."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Get the device date and time."
      }
    },
    "/v1/emark": {
      "post": {
        "description": "Validate a single eMark code. Requires connection: host + password + cashier + pin.",
        "operationId": "emark",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "connection": {
                    "$ref": "#/components/schemas/PartialConn"
                  },
                  "params": {
                    "$ref": "#/components/schemas/SingleEmarkRequest"
                  }
                },
                "required": [
                  "params"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmptyResponse"
                }
              }
            },
            "description": "Operation succeeded."
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Error envelope (4xx/5xx)."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Validate a single eMark code."
      }
    },
    "/v1/header-footer": {
      "post": {
        "description": "Configure receipt header and footer lines. Requires connection: host + password + cashier + pin.",
        "operationId": "headerFooter",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "connection": {
                    "$ref": "#/components/schemas/PartialConn"
                  },
                  "params": {
                    "$ref": "#/components/schemas/SetupHeaderFooterRequest"
                  }
                },
                "required": [
                  "params"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmptyResponse"
                }
              }
            },
            "description": "Operation succeeded."
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Error envelope (4xx/5xx)."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Configure receipt header and footer lines."
      }
    },
...
    "/v1/login": {
      "post": {
        "description": "Verify operator login credentials. Requires connection: host + password + cashier + pin.",
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "connection": {
                    "$ref": "#/components/schemas/PartialConn"
                  },
                  "params": {
                    "description": "Ignored for this operation."
                  }
                },
                "type": "object"
              }
            }
          },
          "required": false
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusOk"
                }
              }
            },
            "description": "Operation succeeded."
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Error envelope (4xx/5xx)."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Verify operator login credentials."
      }
    },
    "/v1/logo": {
      "post": {
        "description": "Configure the receipt header logo. Requires connection: host + password + cashier + pin.",
        "operationId": "headerLogo",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "connection": {
                    "$ref": "#/components/schemas/PartialConn"
                  },
                  "params": {
                    "$ref": "#/components/schemas/SetupHeaderLogoRequest"
                  }
                },
                "required": [
                  "params"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmptyResponse"
                }
              }
            },
            "description": "Operation succeeded."
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Error envelope (4xx/5xx)."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Configure the receipt header logo."
      }
    },
...
    "/v1/operators": {
      "post": {
        "description": "List the HDM's operators and departments. Requires connection: host + password.",
        "operationId": "operators",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "connection": {
                    "$ref": "#/components/schemas/PartialConn"
                  },
                  "params": {
                    "description": "Ignored for this operation."
                  }
                },
                "type": "object"
              }
            }
          },
          "required": false
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListOpsAndDepsResponse"
                }
              }
            },
            "description": "Operation succeeded."
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Error envelope (4xx/5xx)."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "List the HDM's operators and departments."
      }
    },
    "/v1/payment-systems": {
      "post": {
        "description": "List payment systems configured on the device. Requires connection: host + password + cashier + pin.",
        "operationId": "paymentSystems",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "connection": {
                    "$ref": "#/components/schemas/PartialConn"
                  },
                  "params": {
                    "description": "Ignored for this operation."
                  }
                },
                "type": "object"
              }
            }
          },
          "required": false
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentSystemsListResponse"
                }
              }
            },
            "description": "Operation succeeded."
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Error envelope (4xx/5xx)."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "List payment systems configured on the device."
      }
    },
    "/v1/receipt": {
      "post": {
        "description": "Print a fiscal receipt. Requires connection: host + password + cashier + pin.",
        "operationId": "printReceipt",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "connection": {
                    "$ref": "#/components/schemas/PartialConn"
                  },
                  "params": {
                    "$ref": "#/components/schemas/PrintReceiptRequest"
                  }
                },
                "required": [
                  "params"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReceiptResponse"
                }
              }
            },
            "description": "Operation succeeded."
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Error envelope (4xx/5xx)."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Print a fiscal receipt."
      }
    },
    "/v1/receipt/last": {
      "post": {
        "description": "Print a copy of the last receipt. Requires connection: host + password + cashier + pin.",
        "operationId": "printLastReceipt",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "connection": {
                    "$ref": "#/components/schemas/PartialConn"
                  },
                  "params": {
                    "description": "Ignored for this operation."
                  }
                },
                "type": "object"
              }
            }
          },
          "required": false
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmptyResponse"
                }
              }
            },
            "description": "Operation succeeded."
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Error envelope (4xx/5xx)."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Print a copy of the last receipt."
      }
    },
    "/v1/receipt/lookup": {
      "post": {
        "description": "Look up a returnable receipt's contents (read-only). Requires connection: host + password + cashier + pin.",
        "operationId": "lookupReceipt",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "connection": {
                    "$ref": "#/components/schemas/PartialConn"
                  },
                  "params": {
                    "$ref": "#/components/schemas/GetReturnableReceiptRequest"
                  }
                },
                "required": [
                  "params"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReturnableReceiptResponse"
                }
              }
            },
            "description": "Operation succeeded."
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Error envelope (4xx/5xx)."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Look up a returnable receipt's contents (read-only)."
      }
    },
    "/v1/report": {
      "post": {
        "description": "Print an X or Z fiscal report. Requires connection: host + password + cashier + pin.",
        "operationId": "report",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "connection": {
                    "$ref": "#/components/schemas/PartialConn"
                  },
                  "params": {
                    "$ref": "#/components/schemas/FiscalReportRequest"
                  }
                },
                "required": [
                  "params"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmptyResponse"
                }
              }
            },
            "description": "Operation succeeded."
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Error envelope (4xx/5xx)."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Print an X or Z fiscal report."
      }
    },
...
  },
  "servers": [
    {
      "description": "Default local bridge bind address",
      "url": "http://127.0.0.1:8077"
    }
  ]
}